requirements.txt
fastapi
uvicorn[standard]
sqlmodel>=0.0.16
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings>=2.1.0
email-validator>=2.1.0
alembic>=1.13
python-dotenv
pandas
openpyxl